seperation of functionality: ran into some problems
