+++ src/test/londonsw/model/simulation/components/TrafficLightTest.java
@@ -1,7 +1,6 @@
 package londonsw.model.simulation.components;
 
 import londonsw.model.simulation.Ticker;
-import org.junit.Before;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
@@ -18,19 +17,19 @@ public class TrafficLightTest {
 //        ticker = Ticker.getInstance();
 //    }
 
-    @Test
-    public void testNextState() throws Exception {
-        TrafficLight light = new TrafficLight();
-        assertEquals(LightColour.RED, light.state);
-        light.nextState();
-        assertEquals(LightColour.GREEN, light.state);
-        light.nextState();
-//        assertEquals(LightColour.YELLOW, light.state);
+//    @Test
+//    public void testNextState() throws Exception {
+//        TrafficLight light = new TrafficLight();
+//        assertEquals(LightColour.RED, light.state);
 //        light.nextState();
-        assertEquals(LightColour.RED, light.state);
-
-
-    }
+//        assertEquals(LightColour.GREEN, light.state);
+//        light.nextState();
+////        assertEquals(LightColour.YELLOW, light.state);
+////        light.nextState();
+//        assertEquals(LightColour.RED, light.state);
+//
+//
+//    }
 
     @Test
     public void testChange() throws Exception {
